removed unnecessary staging files for galleries, data sourced from pages
