+++ src/pages/comics.js
@@ -1,6 +1,6 @@
 import React from "react"
 import { Link } from "gatsby"
-import ComicsGallery from "../components/page-galleries/comics-gallery"
+import ComicsGallery from "../components/comics-gallery"
 
 import Layout from "../components/layout"
 import SEO from "../components/seo"
