+++ src/pages/index.js
@@ -3,7 +3,7 @@ import React from "react"
 import Layout from "../components/layout"
 
 import SEO from "../components/seo"
-import NavGallery from "../components/page-galleries/nav-gallery"
+import NavGallery from "../components/nav-gallery"
 
 const IndexPage = () => (
   <Layout>
